Change code to use Meteor Accounts package
Heavy revamp of code with the primary focus of migrating existing
functionality to use the Meteor Accounts package and Meteor.user. A lot
of consolidation, restructuring and cleanup happened along the way.

Major changes
- Used accounts-ui for facebook login and removed the existing login
system. Access to FB.api is still retained in client because we need it
to get birthdays.
- Removed all custom collections as Meteor.users is sufficient to
contain all data as of now.
- Moved a lot of functionality to Meteor.methods which is in server.js.
server.js contains full integration with Mongodb which makes it
possible to do more complex queries. This also makes the code more
modular.
- pushed populating friends list out of Template.friends.friends
(because it was obviously not the right place for it to be. This needs
to be moved to the server ideally)

For the meteor facebook api to work, the environment variable ROOT_URL
should be set like this:
ROOT_URL='http://nishanths.info:3000/'

By default meteor sets it to localhost, which doesnt work well with
facebook because it wont match the redirect url set in facebook app
settings. Might be a better idea to manage this using a config file.